refactor(user): 🔨 mover la lógica de negocio a la capa de servicios

+++ BackEnd/src/main/java/com/promanitas/promanitas/services/interfaces/IUserService.java
@@ -8,18 +8,11 @@ import java.util.Optional;
 
 public interface IUserService {
     List<UserEntity> getAllUsers();
-
     UserEntity createUser(UserEntity user);
-
-    Optional<UserEntity> getUserById(Integer userId);
-
-    UserEntity updateUser(Integer userId, UserEntity newUser);
-
-    HashMap<String, String> deleteUser(Integer userId);
-
+    Optional<UserEntity> getUserById(Long userId);
+    UserEntity updateUser(Long userId, UserEntity newUser);
+    HashMap<String, String> deleteUser(Long userId);
     Optional<UserEntity> findByUsername(String username);
-
     Boolean existsByUsername(String username);
-
     Boolean existsByEmail(String email);
 }
